syncronization history

+++ commands/sync/sync.py
@@ -8,19 +8,26 @@ from configs.tokens import youtubeDataApi_token
 
 global playlistID
 playlistID = parameter
+global uploaded_content
+uploaded_content = []
+for x in ReadFF("usr/syncronized.txt").split("\n"):
+    uploaded_content.append(x)
 
 def pageProcess(pageToken):
     global playlistID
+    global uploaded_content
     payload = {"key": youtubeDataApi_token, "maxResults": "5", "playlistId": playlistID, "part": "contentDetails", "pageToken": pageToken}
     headers = {'content-type': 'application/json'}
     r = requests.get('https://youtube.googleapis.com/youtube/v3/playlistItems', params=payload, headers=headers)
     json_data = r.json()
     for video in json_data["items"]:
-        os.system('yt-dlp -f bestvideo+bestaudio --merge-output-format mp4 https://youtube.com/watch?v=' +
-                  video["contentDetails"]["videoId"] + ' --output download.mp4')
-        os.system('telegram-send --video download.mp4 --caption https://youtube.com/watch?v=' + video["contentDetails"][
-            "videoId"])
-        os.system('rm download.mp4')
+        if 'https://youtube.com/watch?v='+video["contentDetails"]["videoId"] not in uploaded_content:
+            os.system('yt-dlp -f bestvideo+bestaudio --merge-output-format mp4 https://youtube.com/watch?v=' +
+                      video["contentDetails"]["videoId"] + ' --output download.mp4')
+            os.system('telegram-send --video download.mp4 --caption https://youtube.com/watch?v=' + video["contentDetails"][
+                "videoId"])
+            os.system('rm download.mp4')
+            PlusWrite('https://youtube.com/watch?v='+video["contentDetails"]["videoId"]+'\n', "usr/syncronized.txt")
 
 pageToken = ""
 payload = {"key": youtubeDataApi_token, "maxResults": "5", "playlistId": playlistID, "part": "contentDetails"}
